\name{ncbiRtools-package}
\alias{ncbiRtools-package}
\alias{ncbiRtools}
\docType{package}
\title{Tools for searching and retrieving sequences in NCBI databases.}
\description{Tools for searching and retrieving sequences in NCBI databases.}
\details{
\tabular{ll}{
Package: \tab ncbiRtools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 19-02-2018\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{Lizandro N. Peraza-Flores

Maintainer: Lizandro N. Peraza-Flores email= lizandropf@gmail.com
}
\references{
For details on database visit:
https://ncbi.nlm.nih.gov
}
\keyword{
  package
  methods
}
